requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-Login==0.6.2
Flask-JWT-Extended==4.5.2
Flask-Migrate==4.0.5
Flask-CORS==4.0.0
SQLAlchemy==2.0.21
Werkzeug==2.3.7
python-dotenv==1.0.0
APScheduler==3.10.4
boto3==1.28.44
botocore==1.31.44
